Add support for variable baudrates
Variable baudrates allows devices in the same bus to work at
different speeds. This way low-bandwidth devices can coexist
with high-bandwidth devices (time and speed permitting) on the
same bus.

This is done by making devices declare a bitmask of supported
BPS and allowing the master to tell them to switch speeds.